apps/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Coffee Culture AI ordering agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-google-genai>=2.0.0",
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-deepgram>=0.9",
    "livekit-plugins-google>=0.9",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.0.0",
]
